+++ src/index.ts
@@ -10,7 +10,7 @@ async function GOOGLE_IMG_SCRAP(config: Config): Promise<Results> {
   verifyGoogleQuery(config);
 
   const URL = constructGoogleUrl(config);
-  const result = await parseGoogleImages(URL);
+  const result = await parseGoogleImages(URL, config.proxy);
   const slicedResult = limitResultSize(config?.limit, result);
 
   return {